% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{adjacent.coords.plot}
\alias{adjacent.coords.plot}
\title{Plot adjacent gene pair coordinates}
\usage{
adjacent.coords.plot(adjacent.coords = NULL, pair = NULL,
  bw.plus = NULL, bw.minus = NULL, bp.bin = 5, xper = 0.2,
  yper = 1.3)
}
\arguments{
\item{adjacent.coords}{bed6 frame with coordinates for adjacent genes}

\item{pair}{frame with one row, upstream and downstream genes}

\item{bw.plus}{plus strand bigWig data}

\item{bw.minus}{minus strand bigWig data}

\item{xper}{fraction of the x-axis before the gene starts}

\item{yper}{fraction of the y-axis with open space above the read data}
}
\value{
A plot with read data and gene coordinate annotations.
}
\description{
Plot the coordinates, along with read data, for overlapping genes that were identified as an adjacent pair.
}
\examples{
# visualize coordinates for a specific pair
adjacent.coords.plot(adjacent.coords=adjacent.coords,
                  pair=fix.genes[4,],
                  bw.plus=bw.plus, bw.minus=bw.minus)

}
