% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{adjacent.gene.coords}
\alias{adjacent.gene.coords}
\title{Function to identify coordinates for adjacent gene pairs}
\usage{
adjacent.gene.coords(fix.genes = NULL, bed.long = NULL, exon1 = NULL,
  bw.plus = NULL, bw.minus = NULL, bp.bin = NULL, shift.up = NULL,
  dist.from.start = 50, delta.tss = NULL, knot.div = 4,
  knot.thresh = 5, diff.tss = 1000, pause.bp = 120,
  fname = "adjacentSplines.pdf")
}
\arguments{
\item{fix.genes}{frame with upstream and downstream genes}

\item{bed.long}{long gene annotations, see get.largest.interval()}

\item{exon1}{bed6 frame with first exon gene annotations}

\item{bw.plus}{plus strand bigWig data}

\item{bw.minus}{minus strand bigWig data}

\item{bp.bin}{the interval will be separated into adjacent bins of this size}

\item{shift.up}{look upstream of the long gene start by this amount to search for a viable TSS}

\item{dist.from.start}{distance between the TTS of the upstream gene and the TSS of the downstream gene}

\item{delta.tss}{amount by which to shift the identied TSS upstream of the identified interval}

\item{knot.div}{the number of knots for the spline fit is defined as the number of bins covering the gene end divided by this number}

\item{knot.thresh}{minimum number of knots}

\item{diff.tss}{minimum distance between TSSs}

\item{pause.bp}{number of bp by which a pause site should be considered downstrean of a TSS}

\item{fname}{file name (.pdf) for output plots}
}
\value{
A frame with the adjusted coordinates for the input gene set, along with a plot. The titles for the plots indicate the upstream gene, the downstream gene, and the strand.
For genes on the plus strand, the upstream gene is on the left. For the minus strand, the upstream gene is on the right.
The red line denotes the spline fit and the vertical lines indicate the TSSs.
The plot is intended for diagnostic purposes.
}
\description{
We have identified adjacent gene pairs with overlaps based on manual analyses.
We empirically define TSSs for these genes by binning the region spanned by both genes,
fitting smooth spline curves to the binned read counts, and identifying the two largest peaks separated
by a specified distance. We set a bin size and apply the constraint that the identified
'pause' peaks must be a given distance apart. The search region includes a specified distance upstream of
the upstream-most gene. We shift the identified peaks upstream. For the spline fits, we set the
number of knots to the number of bins divided by specified setting.
We identify the TSSs as the exon 1 coordinates closest to the pasue site peaks.
We identify the TTS of the upstream gene as an interval from the TSS of the downstream gene.
The input TTS for the downstream gene is retained.
}
\examples{
# get the start sites for adjacent gene pairs
fix.genes = rbind(fix.genes.id, fix.genes.ov)
bp.bin = 5
knot.div = 40
shift.up = 100
delta.tss = 50
diff.tss = 1000
dist.from.start  = 50
adjacent.coords = adjacent.gene.coords(fix.genes=fix.genes, bed.long=TSS.gene,
                                       exon1=gencode.firstExon,
                                       bw.plus=bw.plus, bw.minus=bw.minus,
                                       knot.div=knot.div, bp.bin=bp.bin,
                                       shift.up=shift.up, delta.tss=delta.tss,
                                       dist.from.start=dist.from.start,
                                       diff.tss=diff.tss, fname="adjacentSplines.pdf")

}
