% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{adjacent.gene.tts}
\alias{adjacent.gene.tts}
\title{Function to identify gene ends (TTSs) for adjacent gene pairs}
\usage{
adjacent.gene.tts(fix.genes = NULL, bed.long = NULL,
  TSS.adjacent = NULL, dist.from.start = NULL)
}
\arguments{
\item{fix.genes}{frame with upstream and downstream genes for adjacent gene pairs}

\item{bed.long}{comprehensive gene annotations with large intervals defined by get.largest.interval()}

\item{TSS.adjacent}{TSSs for adjancent genes, identified from adjacent.gene.tss()}

\item{dist.from.start}{distance between the TTS of the upstream gene and the TSS of the downstream gene}
}
\value{
A bed6 frame with TTSs for upstream genes in adjacent gene pairs
}
\description{
We set the TTSs for the upstream genes to a given distance from the starts of the downstream genes for the
adjacent gene pairs. See also adjacent.gene.tss().
}
\examples{
# get the end sites for adjacent upstream genes, integrate with TSSs
dist.from.start = 50
adjacent.tts.up = adjacent.gene.tts(fix.genes=fix.genes, bed.long=bed.long,
                                 TSS.adjacent=TSS.adjacent, dist.from.start=dist.from.start)
tss = TSS.adjacent$TSS
names(tss) = TSS.adjacent$gene
adjacent.tts.up = apply.TSS.coords(bed=adjacent.tts.up, tss=tss)
}
