% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{apply.TSS.coords}
\alias{apply.TSS.coords}
\title{Apply previously identified TSSs to an existing gene annotation}
\usage{
apply.TSS.coords(bed = NULL, tss = NULL)
}
\arguments{
\item{bed}{A bed6 frame with gene annotations}

\item{tss}{TSS estimates, from inside get.TSS().
Note that the TSS genes must be a subset of the genes in the bed6 data.}
}
\value{
A bed6 frame with estimated TSSs incorporated
}
\description{
This function will apply transcription start site (TSS) coordinates to a bed6 frame.
This function is embedded within get.TSS().
}
\examples{
out = apply.TSS.coords(bed=bed.out, tss=tss)
}
