% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tu.R
\name{assess.final.coords}
\alias{assess.final.coords}
\title{Function to assessing the final identified coordinates}
\usage{
assess.final.coords(bed1 = NULL, bed2 = NULL)
}
\arguments{
\item{bed1}{A reference bed frame from get TTS()}

\item{bed2}{A final bed frame from tu.gene.overlaps()}
}
\value{
A list of metrics: dtss is a vector of tss differences,
dtts is a vector of tts differences,
dtss.mean is the average tss differences,
dtss.uneq is the number of tss's that are not identical
dtts.mean is the mean for tts differences,
and dtts.uneq is the number of non-identical tts's
}
\description{
This function compares the final coordinates to coordinates to inferred coordinates,
tu.gene.overlaps() versus get.TTS().
}
\examples{
# evaluate the final annotation
metrics = assess.final.coords(bed1=bed.tss.tts, bed2=res)
metrics$dtss.uneq
metrics$dtts.uneq

}
