% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{chrom.size.filter}
\alias{chrom.size.filter}
\title{Filter identified gene coordinates for chromosome sizes.}
\usage{
chrom.size.filter(coords = NULL, chrom.sizes = NULL)
}
\arguments{
\item{coords}{bed6 frame with inferred coordinates}

\item{chrom.sizes}{two column file with chromosomes and their respective sizes in bp}
}
\value{
A list with a bed6 frame with corrected coordinates (bed)
and a log documenting which corrections, if any, were made (log).
}
\description{
First acquire a chrom.sizes file (e.g., http://hgdownload.cse.ucsc.edu/goldenPath/mm10/bigZips/mm10.chrom.sizes; see vignette).
If any ends are greater than the species chromosome size, reduce the end to the size limit.
Any negative starts will be set to 0.
}
\examples{
coords.filt = chrom.size.filter(coords=coords, chrom.sizes=chrom.sizes)
head(coords.filt$bed)
coords.filt$log

}
