% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{gene.overlaps}
\alias{gene.overlaps}
\title{Documentation of gene overlaps}
\usage{
gene.overlaps(bed = NULL)
}
\arguments{
\item{bed}{bed6 file with processed gene annotations}
}
\value{
A list with has.start.inside and is.a.start.inside. has.start.inside = bed6 file that documents genes in which there are starts
        of other genes. is.a.start.inside = bed6 file that documents genes that start inside other
        genes.
}
\description{
This function identifies gene overlaps that do not involve identical coordinates.
}
\examples{
# run overlap analysis
overlap.data = gene.overlaps( bed = bed.long.filtered2.tss )
has.start.inside = overlap.data$has.start.inside
is.a.start.inside = overlap.data$is.a.start.inside
dim(has.start.inside)
dim(is.a.start.inside)
}
