% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{get.dups}
\alias{get.dups}
\title{Get duplicates in which start sites match or end sites match}
\usage{
get.dups(bed = NULL)
}
\arguments{
\item{bed}{a bed file}
}
\value{
A bed file with duplicate entries. Col 7 indicates ordered overlap loci.
}
\description{
This function identifies an extreme for of gene overlaps in which multiple identifiers
correspond to identical start or end coordinates.
}
\examples{
dups0 = get.dups(bed = TSS.gene)

}
