% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{get.largest.interval}
\alias{get.largest.interval}
\title{Get the largest interval for each gene, given multiple TSS and TTS annotations}
\usage{
get.largest.interval(bed = NULL)
}
\arguments{
\item{bed}{bed6 frame with comprehensive gene annotations, defaults to NULL}
}
\value{
A bed6 frame with the largest annotation for each gene
}
\description{
The input bed6 file can be derived from a gencode annotation file,
as described in the vignette
}
\examples{
# get intervals for furthest TSS and TTS +/- interval
bed.long = get.largest.interval(bed=dat0)
}
