% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{get.TSS}
\alias{get.TSS}
\title{Select an optimal transcription start site (TSS) for each gene}
\usage{
get.TSS(bed.in = NULL, bed.out = NULL, bw.plus = NULL,
  bw.minus = NULL, bp.range = NULL, cnt.thresh = NULL,
  low.read = FALSE)
}
\arguments{
\item{bed.in}{bed6 frame with first exon gene annotations}

\item{bed.out}{bed6 frame for output (same genes as bed.in)}

\item{bw.plus}{plus strand bigWig data}

\item{bw.minus}{minus strand bigWig data}

\item{bp.range}{range in bp for identifying the exon 1 with the greatest density}

\item{cnt.thresh}{read count number below which the TSS is not evaluated, by default the gene is removed from the analysis max(counts) < cnt.thresh.
Indicate low.read=TRUE to use the upstream-most TSS for max(counts) < cnt.thresh.}

\item{low.read}{Logical for whether to pick the upstream-most TSS if max(counts) < cnt.thresh (default FALSE)}
}
\value{
A list with a bed6 frame with inferred TSSs (bed) and a list of problematic genes (issues).
Problems documented include start > end and start < 0. For start > end,
the original input corrdinates were retained (bed.out).
For start < 0, start = 0 was applied.
}
\description{
We set a range around each annotated exon 1 within which to search for a read density.
For each gene, we select the upstream coordinate of the first exon with the greatest read density, in the specified region,
out of all first exons annotated for the given gene.
The output of this function should be processed using the function apply.TSS.coords().
Note that input and output genes must be matched.
}
\examples{
# select the TSS for each gene and incorporate these TSSs into the largest interval coordinates
bp.range = c(20,120)
cnt.thresh = 5
bed.out = largest.interval.expr.bed
bed.in = gencode.firstExon[gencode.firstExon$gene \%in\% bed.out$gene,]
TSS.gene = get.TSS(bed.in=bed.in, bed.out=bed.out,
                   bw.plus=bw.plus, bw.minus=bw.minus,
                   bp.range=bp.range, cnt.thresh=cnt.thresh)
}
