% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{get.TTS}
\alias{get.TTS}
\title{Estimate the transcription termination sites (TTSs)}
\usage{
get.TTS(bed = NULL, bw = NULL, bp.bin = NULL, pk.thresh = NULL,
  knot.div = NULL, cnt.thresh = NULL, knot.thresh = NULL,
  add.to.end = NULL, tau.dist = NULL, frac.max = NULL,
  frac.min = NULL)
}
\arguments{
\item{bed}{bed6 gene coordinates with gene end intervals for estimation of the TTS}

\item{bw}{plus or minus strand bigWig data}

\item{bp.bin}{the interval at the gene end will be separated into adjacent bins of this size}

\item{pk.thresh}{the TTS is defined as pk.thresh percent of max peak of the spline fit}

\item{knot.div}{the number of knots for the spline fit is defined as the number of bins covering the gene end divided by this number}

\item{cnt.thresh}{read count number below which the TTS is not evaluated}

\item{knot.thresh}{minimum number of knots, if knot.div gives a lower number, use this}

\item{add.to.end}{the maximal length of the search region (bp)}

\item{tau.dist}{distance constant for the exponential defining the region for peak detection}

\item{frac.max}{maximal fraction of gene end region for peal detection}

\item{frac.min}{minimal fraction of gene end region for peal detection}
}
\value{
A vector of TTS coordinates
}
\description{
This function estimates the TTSs, this function is called by get.gene.end() and is not designed to be run on its own.
}
\examples{
See get.gene.end()
}
