% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{inside.starts}
\alias{inside.starts}
\title{Identify genes with multiple starts inside}
\usage{
inside.starts(vec = NULL)
}
\arguments{
\item{vec}{input the xy column from the $is.a.start.inside output of gene.overlaps()}
}
\value{
A vector of genes within which multiple starts are observed.
}
\description{
This function identifies genes within which >1 start is observed.
That is, relatively 'large' genes in which 2 or more genes originate.
Data generated from gene.overlaps() serve as input to this function.
}
\examples{
overlap.data = gene.overlaps( bed = TSS.gene.filtered1 )
is.a.start.inside = overlap.data$is.a.start.inside
mult.inside.starts = inside.starts(vec = is.a.start.inside$xy)
}
