% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tu.R
\name{multi.overlap.assign}
\alias{multi.overlap.assign}
\title{Function to assign identifiers for class5 overlap TUs}
\usage{
multi.overlap.assign(fr = NULL, tss.thresh = NULL, delta.tss = NULL,
  delta.tts = NULL, cl = NULL)
}
\arguments{
\item{fr}{= frame with full bedtools overlap data for a class5 TU
with multiple internal annotations}

\item{tss.thresh}{= number of bp a TU beginning can be off
from an annotation in order to be assigned that annotation}

\item{delta.tss}{= max distance between an upstream gene end and
downstream gene start cl = multi-overlap class}

\item{delta.tts}{max difference distance between and annotated gene
end and the start of a downstream gene before an intermediate TU id is assigned}

\item{cl}{multi-overlap class}
}
\value{
A bed data with chr, start, end, id, class, and strand.
Note that id = 0 for regions of large TUs that do no match input annotations.
}
\description{
This function is called inside of multiple.overlaps()
and is not recommended to be used on its own.
}
\examples{
See multiple.overlaps()
}
