% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tu.R
\name{multi.overlaps}
\alias{multi.overlaps}
\title{Assign identifiers to TUs with multiple gene overlaps}
\usage{
multi.overlaps(overlaps = NULL, tss.thresh = NULL, delta.tss = NULL,
  delta.tts = NULL)
}
\arguments{
\item{overlaps}{A frame with bed intersect of inferred TUs (col1-6) with annotated TUs (col7-12) and overlaping bps (col14)}

\item{tss.thresh}{Number of bp a TU beginning can be off from an annotation in order to be assigned that annotation}

\item{delta.tss}{Max distance between an upstream gene end and downstream gene start}

\item{delta.tts}{Max difference distance between and annotated gene end and the start of a downstream gene before an intermediate TU id is assigned}
}
\value{
A list with bed data (bed) and counts for each class (cnt5-8).
bed = data with chr, start, end, id, class, and strand.
cnt5 = count of TU from class 5 overlaps
}
\description{
This function will assign identifiers to TUs that overlapped with multiple genes.
Trusted gene annotations are considered in terms of their degree of overlap with
TUs identified in an unbiased manner. Marginal regions of TUs outside of gene overlaps
are given generic identifiers.
}
\examples{
dup.hmm.rows = overlap.tu[duplicated(overlap.tu[,c(1:3,6)]) |
    duplicated(overlap.tu[,c(1:3,6)], fromLast=TRUE),]
dup.full = dup.hmm.rows
names(dup.full)[1:6] = c("infr.chr","infr.start","infr.end",
                         "infr.gene","infr.xy","infr.strand")
nrow(dup.full[,1:3] \%>\% unique)
nrow(dup.full \%>\% select(ann.gene) \%>\% unique)
tss.thresh = 200
delta.tss = 50
delta.tts = 1000
class5678 = multi.overlaps(overlaps=dup.full, tss.thresh=tss.thresh,
                           delta.tss=delta.tss, delta.tts=delta.tts)
new.ann.mult = class5678$bed
}
