% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{read.count.body}
\alias{read.count.body}
\title{Count gene body reads and compute density}
\usage{
read.count.body(bed = NULL, bw.plus = NULL, bw.minus = NULL,
  bp.start = 0)
}
\arguments{
\item{bed}{a bed6 frame}

\item{bw.plus}{plus strand bigWig data}

\item{bw.minus}{minus strand bigWig data}

\item{bp.start}{optional number of bases to omit from the upstream gene boundary}
}
\value{
A list with counts and desity.
counts = vector with end of gene counts for each gene.
density = vector with end of gene densities for each gene.
}
\description{
Input coordinates for evaluation of counts and densities.
This function works with input that includes one annotation per gene.
}
\examples{
# get read counts and densities at the body of each annotated gene

body.reads = read.count.body(bed=largest.interval.bed, bw.plus=bw.plus,
bw.minus=bw.minus, bp.start=bp.start)
hist(log(body.reads$density), breaks=200, col="black",xlab="log read density",main="")
hist(log(body.reads$counts), breaks=200, col="black",xlab="log read count",main="")
}
