% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{read.count.end}
\alias{read.count.end}
\title{Select a regions containing the gene ends}
\usage{
read.count.end(bed = NULL, bw.plus = NULL, bw.minus = NULL,
  fraction.end = NULL, add.to.end = NULL)
}
\arguments{
\item{bed}{a bed6 frame}

\item{bw.plus}{plus strand bigWig data}

\item{bw.minus}{minus strand bigWig data}

\item{fraction.end}{fraction of the annotated gene end (0,1)}

\item{add.to.end}{number of base pairs beyond the gene end within which to count reads}
}
\value{
A list with counts and desity.
counts = vector with end of gene counts for each gene.
density = vector with end of gene densities for each gene.
}
\description{
Select a fraction of the annotated gene end and consider an additional number of base pairs beyond
the gene end within which to count reads.
This function works with input that includes one annotation per gene.
}
\examples{
# get read counts and densities at the end of each annotated gene
fraction.end = 0.1
add.to.end = 0
end.reads = read.count.end(bed=bed.long, bw.plus=bw.plus, bw.minus=bw.minus,
                          fraction.end=fraction.end, add.to.end=add.to.end)
hist(log(end.reads$density))
hist(log(end.reads$counts))
}
