% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{read.count.transcript}
\alias{read.count.transcript}
\title{Evaluate reads and compute densities for each transcript of each gene.}
\usage{
read.count.transcript(bed = NULL, bw.plus = NULL, bw.minus = NULL)
}
\arguments{
\item{bed}{a bed6 frame}

\item{bw.plus}{plus strand bigWig data}

\item{bw.minus}{minus strand bigWig data}
}
\value{
A list with counts and desity.
counts = vector with end of gene counts for each gene.
density = vector with end of gene densities for each gene.
}
\description{
Input coordinates for evaluation of counts and densities.
Return the maximal values across all transcripts of a given gene.
In particular, we return the density associated with the max count for a specific gene.
}
\examples{
# get read counts and densities for the max transcript of each annotated gene
transcript.reads = read.count.transcript(bed=gencode.transcript, bw.plus=bw.plus, bw.minus=bw.minus)
hist(log(transcript.reads$density), breaks=200, col="black",xlab="log read density",main="")
hist(log(transcript.reads$counts), breaks=200, col="black",xlab="log read count",main="")
}
