% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tu.R
\name{single.overlap.assign}
\alias{single.overlap.assign}
\title{Function to assign identifiers for class1 overlap TUs}
\usage{
single.overlap.assign(fr = NULL, tss.thresh = NULL, delta.tss = NULL,
  delta.tts = NULL, cl = NULL)
}
\arguments{
\item{fr}{A frame with full bedtools overlap data for a class1}

\item{tss.thresh}{Number of bp a TU beginning can be off
from an annotation in order to be assigned that annotation}

\item{delta.tss}{Max distance between an upstream gene end and downstream gene start}

\item{cl}{Overlap class}
}
\value{
A bed data frame with chr, start, end, id, class, and strand.
Note that id = 0 for regions of large TUs that do no match input annotations.
}
\description{
This function is called inside of single.overlaps()
and is not recommended to be used on its own.
}
\examples{
See single.overlaps()
}
