% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tu.R
\name{single.overlaps}
\alias{single.overlaps}
\title{Assign identifiers to TUs with single gene overlaps}
\usage{
single.overlaps(overlaps = NULL, tss.thresh = NULL, delta.tss = NULL,
  delta.tts = NULL)
}
\arguments{
\item{overlaps}{A frame with bed intersect of inferred TUs (col1-6)
with annotated TUs (col7-12) and overlaping bps (col14)}

\item{tss.thresh}{The number of bp that a TU beginning can be off from an annotation in
order to be assigned that annotation}

\item{delta.tss}{Max distance between an upstream gene end and downstream gene start}

\item{delta.tts}{Max difference distance between and annotated gene end and the start
of a downstream gene before an intermediate TU id is assigned}
}
\value{
A list with bed data (bed) and counts for each class (cnt1-4).
bed = data with chr, start, end, id, class, and strand.
cnt1 = count of TU from class 1 overlaps
}
\description{
This function will assign identifiers to TUs that overlapped with single genes.
Trusted gene annotations are considered in terms of their degree of overlap with
TUs identified in an unbiased manner. Marginal regions of TUs outside of gene overlaps
are given generic identifiers.
}
\examples{
sing.hmm.rows = overlap.tu[!duplicated(overlap.tu[,c(1:3,6)]) &
     !duplicated(overlap.tu[,c(1:3,6)], fromLast=TRUE),]
overlaps = sing.hmm.rows
names(overlaps)[1:6] = c("infr.chr","infr.start","infr.end",
                         "infr.gene","infr.xy","infr.strand")
tss.thresh = 200
delta.tss = 50
delta.tts = 1000
class1234 = single.overlaps(overlaps=overlaps, tss.thresh=tss.thresh,
                            delta.tss=delta.tss, delta.tts=delta.tts)
new.ann.sing = class1234$bed
}
