% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{TTS.boundry.clip}
\alias{TTS.boundry.clip}
\title{Look at whether TTSs identified at the search boundry were clipped.}
\usage{
TTS.boundry.clip(coords = NULL, bed.for.tts.eval = NULL)
}
\arguments{
\item{coords}{A bed6 frame with inferred coordinates (from get.TTS())}

\item{bed.for.tts.eval}{A bed6 frame with TTS search regions (from get.end.intervals())}
}
\value{
A list with two fractions [0,1], frac.clip and frac.noclip,
and a vector of clip distances (clip.tts) for genes with boundary TTSs.
}
\description{
This function considers information in both coords and bed.for.tts.eval (see vignette)
to identify the fractions of search boundary TTSs that were clipped and unclipped.
}
\examples{
# look at whether TTSs identified at the search boundry had clipped boundries
frac.bound.clip = TTS.boundry.clip(coords=coords, bed.for.tts.eval=bed.for.tts.eval)
frac.bound.clip$frac.clip
frac.bound.clip$frac.noclip

# plot didtribution of clip distances for genes with boundary TTSs
hist(frac.bound.clip$clip.tts,main="",xlab="bp clipped for boundry genes")

}
