% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{tts.plot}
\alias{tts.plot}
\title{Plot inferred and largest interval coordinates}
\usage{
tts.plot(coords = NULL, gene.end = NULL, long.gene = NULL,
  bw.plus = NULL, bw.minus = NULL, bp.bin = 5, gene = NULL,
  xper = 0.2, yper = 1.3, add.to.end = NULL, tau.dist = NULL,
  frac.max = NULL, frac.min = NULL)
}
\arguments{
\item{coords}{bed6 frame with inferred coordinates}

\item{gene.end}{bed6 frame with peak search regions}

\item{long.gene}{bed6 frame with largest interval coordinates}

\item{bw.plus}{plus strand bigWig data}

\item{bw.minus}{minus strand bigWig data}

\item{gene}{gene for plotting}

\item{xper}{fraction of the x-axis before the gene starts}

\item{yper}{fraction of the y-axis with open space above the read data}

\item{add.to.end}{the maximal length of the search region (bp)}

\item{tau.dist}{distance constant for the exponential defining the region for peak detection}

\item{frac.max}{maximal fraction of gene end region for peal detection}

\item{frac.min}{minimal fraction of gene end region for peal detection}
}
\value{
A plot with reads, interred annptation (black) largest interval annotation (gray),
TTS search interval (solid, vertical), and TTS peak interval (dashed, vertical).
}
\description{
Plot the inferred and largest interval coordinates, along with read data,
indicate the peak search region, and indicate the total search region.
}
\examples{
gene.end = bed.for.tts.eval
long.gene = largest.interval.bed
tts.plot(coords=coords, gene.end=gene.end, long.gene=long.gene,
         gene = "Pparg", xper=0.2, yper=0.3,
         frac.min=frac.min, frac.max=frac.max,
         bw.plus=bw.plus, bw.minus=bw.minus, bp.bin=5,
         add.to.end=add.to.end, tau.dist=tau.dist)

}
