% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_tu.R
\name{tu.gene.overlaps}
\alias{tu.gene.overlaps}
\title{Assign identifiers to TUs with single gene overlaps}
\usage{
tu.gene.overlaps(hmm.ann.overlap = NULL)
}
\arguments{
\item{hmm.ann.overlap}{A frame with coordinates from intersecting tus with inferred gene annotations}
}
\value{
A list of metrics inclusing the total number of TUs (n.tus),
the number of TUs without gene overlaps (n.tu.no.gene.overlap),
the number of TUs with overlapping genes (n.tu.overlap.gene),
the number of genes with overlapping TUs (n.gene.overlap.tu),
the number of TUs overlapping single genes (n.tu.overlap.single.genes),
and the number of TUs with multiple gene overlaps (n.tu.overlap.multiple.genes).
Note that n.tu.overlap.gene = n.tu.overlap.single.genes + n.tu.overlap.multiple.genes.
}
\description{
Summarize gene/tu counts after intersecting the respective annotations.
See the vignette for analysis details and instructions.
}
\examples{
tg.overlaps = tu.gene.overlaps(hmm.ann.overlap=hmm.ann.overlap)

}
