% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{adjacent.coords.plot}
\alias{adjacent.coords.plot}
\title{Plot adjacent gene pair coordinates}
\usage{
adjacent.coords.plot(adjacent.coords = NULL, pair = NULL,
  bw.plus = NULL, bw.minus = NULL, bp.bin = 5, xper = 0.2,
  yper = 0.3)
}
\arguments{
\item{adjacent.coords}{A bed6 frame with coordinates for adjacent genes}

\item{pair}{Frame with one row, upstream and downstream genes}

\item{bw.plus}{Plus strand bigWig data}

\item{bw.minus}{Minus strand bigWig data}

\item{xper}{Fraction of the x-axis before the gene starts}

\item{yper}{Fraction of the y-axis with open space above the read data}
}
\value{
A plot with read data and gene coordinate annotations.
}
\description{
Plot the coordinates, along with read data, for overlapping genes
that were identified as an adjacent pair,
with coordinates determined using adjacent.gene.coords().
}
\examples{
# visualize coordinates for a specific pair
adjacent.coords.plot(adjacent.coords=adjacent.coords,
                  pair=fix.genes[4,],
                  bw.plus=bw.plus, bw.minus=bw.minus)

}
