% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{apply.TSS.coords}
\alias{apply.TSS.coords}
\title{Apply previously identified TSSs to an existing gene annotation}
\usage{
apply.TSS.coords(bed = NULL, tss = NULL)
}
\arguments{
\item{bed}{bed6 file with comprehensive gene annotations}

\item{tss}{TSS estimates, output from get.TSS().
Note that the TSS genes must be a subset of the genes in the bed6 data.}
}
\value{
A bed6 file with estimated TSSs incorporated
}
\description{
This function will apply transcription start site (TSS) coordinates to a bed6 frame to incorporate the output from
get.TSS() into an existing bed frame.
}
\examples{
# apply TSS coordinates to the expression-filtered long gene annotations
bed.long.filtered.tss = apply.TSS.coords(bed=bed.long.filtered, tss=TSS.gene)
}
