% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{gene.end.plot}
\alias{gene.end.plot}
\title{Plot the results of transcription terminination site identification}
\usage{
gene.end.plot(bed = NULL, gene = NULL, bw.plus = NULL,
  bw.minus = NULL, bp.bin = NULL, pk.thresh = 0.1, knot.div = 40,
  knot.thresh = 5, cnt.thresh = 5, add.to.end = 50000,
  tau.dist = 10000, frac.max = 1, frac.min = 0.2)
}
\arguments{
\item{bed}{Bed6 gene coordinates with gene end intervals for estimation of the TTS}

\item{gene}{A specific gene for plotting}

\item{bw.plus}{Plus strand bigWig data}

\item{bw.minus}{Minus strand bigWig data}

\item{bp.bin}{The interval at the gene end will be separated into adjacent bins of this size}

\item{pk.thresh}{The TTS is defined as pk.thresh percent of max peak of the spline fit}

\item{knot.div}{The number of knots for the spline fit is defined as the number of bins covering the gene end divided by this number;
increasing this parameter results and a smoother curve}

\item{knot.thresh}{Minimum number of knots, if knot.div gives a lower number, use this}

\item{cnt.thresh}{Read count number below which the TTS is not evaluated}

\item{add.to.end}{The maximal length of the search region (bp)}

\item{tau.dist}{Distance constant for the exponential defining the region for peak detection}

\item{frac.max}{Maximal fraction of gene end region for peal detection}

\item{frac.min}{Minimal fraction of gene end region for peal detection}
}
\value{
A single plot is generated with the binned reads (circles), a fitted curve (red),
and the location of the identified TTS (vertical line).
}
\description{
This function can be used to evaluate the performance of get.TTS().
We also recommend visualizing the results of data-driven gene annotations using a genome browser.
}
\examples{
gene.end.plot(bed=bed.for.tts.eval, gene="Pparg",
              bw.plus=bw.plus, bw.minus=bw.minus,
              bp.bin=bp.bin, add.to.end=add.to.end, knot.div=knot.div,
              pk.thresh=pk.thresh, knot.thresh=knot.thresh,
              cnt.thresh=cnt.thresh, tau.dist=tau.dist,
              frac.max=frac.max, frac.min=frac.min)
}
