% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{gene.overlaps}
\alias{gene.overlaps}
\title{Documentation of gene overlaps}
\usage{
gene.overlaps(bed = NULL)
}
\arguments{
\item{bed}{A bed6 frame with processed gene annotations}
}
\value{
A list with has.start.inside, is.a.start.inside, and cases. The object has.start.inside is a bed6 frame that documents genes in which there are starts
of other genes. The object is.a.start.inside is a bed6 frame that documents genes that start inside other genes. The object
cases documents loci in which 2 or more genes overlap (col 7 of a bed format frame).
}
\description{
This function identifies gene overlaps.
}
\examples{
# run overlap analysis
overlap.data = gene.overlaps( bed = TSS.gene.filtered1 )
has.start.inside = overlap.data$has.start.inside
is.a.start.inside = overlap.data$is.a.start.inside
head(has.start.inside)
head(is.a.start.inside)
head(overlap.data$cases)
length(unique(overlap.data$cases$gene))

}
