% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{get.gene.end}
\alias{get.gene.end}
\title{Estimate the transcription termination sites (TTSs)}
\usage{
get.gene.end(bed = NULL, bw = NULL, bp.bin = NULL,
  pk.thresh = NULL, knot.div = NULL, cnt.thresh = NULL,
  knot.thresh = NULL, add.to.end = NULL, tau.dist = NULL,
  frac.max = NULL, frac.min = NULL, sum.thresh = NULL)
}
\arguments{
\item{bed}{A bed6 gene coordinates with gene end intervals for estimation of the TTS}

\item{bw}{Plus or minus strand bigWig data}

\item{bp.bin}{The interval at the gene end will be separated into adjacent bins of this size}

\item{pk.thresh}{The TTS is defined as pk.thresh percent of max peak of the spline fit (identified within the peak search region)}

\item{knot.div}{The number of knots for the spline fit is defined as the number of bins covering the gene end divided by this number}

\item{cnt.thresh}{Read count number below which the TTS is not evaluated}

\item{knot.thresh}{Minimum number of knots, if knot.div gives a lower number, use this}

\item{add.to.end}{The maximal length of the search region (bp)}

\item{tau.dist}{Distance constant for the exponential function defining the region for peak detection}

\item{frac.max}{Maximal fraction of gene end region for peak detection}

\item{frac.min}{Minimal fraction of gene end region for peak detection}

\item{sum.thresh}{Threshold for the min sum of read counts}
}
\value{
A vector of TTS coordinates
}
\description{
This function estimates the TTSs. This function is called by get.TTS() 
and is not designed to be run on its own. TTSs that cannot be estimated are assigned zeros.
}
\examples{
See get.TTS()
}
