% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{get.TSS}
\alias{get.TSS}
\title{Select an optimal transcription start site (TSS) for each gene}
\usage{
get.TSS(bed = NULL, bw.plus = NULL, bw.minus = NULL,
  bp.range = NULL, bp.delta = NULL, bp.bin = NULL,
  delta.tss = NULL, cnt.thresh = NULL)
}
\arguments{
\item{bed}{bed6 file with comprehensive gene annotations}

\item{bw.plus}{plus strand bigWig data}

\item{bw.minus}{minus strand bigWig data}

\item{bp.range}{range to be centered on each gene start for evaluating read counts.
Note that the value should be divisible by 2}

\item{bp.delta}{number of reads to move incrementally - sliding window interval}

\item{bp.bin}{bin size for evaluating read counts within a given region}

\item{delta.tss}{amount by which to shift the identied TSS upstream of the identified interval}

\item{cnt.thresh}{read count number below which the TTS is not evaluated}
}
\value{
A vector of indentified transcription start site coordinates. Note that strand and chromosome information are not provided here.
}
\description{
We set a range around each annotated gene start within which to search for a region of peak read
density. Such regions of peak read density occur at ’pause sites’ that are typically 20-80 bp from
the TSS. Within each region around an annotated gene start, we translate a sliding
window throughout to find the sub-region with maximal density. We determine the window with the
maximal density out of all regions considered and define the TSS as the upstream boundry of this
window, minus a shift of a specified amount. The output of this function should be processed using the function apply.TSS.coords().
}
\examples{
# select the TSS for each gene
bp.range = 1200
bp.delta = 10
bp.bin = 50
delta.tss = 50
cnt.thresh = 5
TSS.gene = get.TSS(bed=dat0, bw.plus=bw.plus, bw.minus=bw.minus,
                  bp.range=bp.range, bp.delta=bp.delta, bp.bin=bp.bin,
                  delta.tss=delta.tss, cnt.thresh=cnt.thresh)
}
