% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{get.TTS}
\alias{get.TTS}
\title{Defining gene ends}
\usage{
get.TTS(bed = NULL, tss = NULL, bw.plus = NULL, bw.minus = NULL,
  bp.bin = NULL, pk.thresh = 0.1, knot.div = 40, knot.thresh = 5,
  cnt.thresh = 5, add.to.end = 50000, sum.thresh = 20,
  tau.dist = 10000, frac.max = 1, frac.min = 0.2)
}
\arguments{
\item{bed}{A bed6 gene coordinates with gene end intervals for estimation of the TTS}

\item{tss}{A bed6 gene coordinates containing estimated TSSs and largest interval gene ends}

\item{bw.plus}{Plus strand bigWig data}

\item{bw.minus}{Minus strand bigWig data}

\item{bp.bin}{The interval at the gene end will be separated into adjacent bins of this size}

\item{pk.thresh}{The TTS is defined as pk.thresh percent of max peak of the spline fit}

\item{knot.div}{the number of knots for the spline fit is defined as the number of bins covering the gene end divided by this number;
increasing this parameter results and a smoother curve}

\item{knot.thresh}{Minimum number of knots, if knot.div gives a lower number, use this}

\item{cnt.thresh}{Read count number below which the TTS is not evaluated}

\item{add.to.end}{The maximal length of the search region (bp)}

\item{sum.thresh}{Threshold for the min sum of read counts, below this default to the input TTS (default 20)}

\item{tau.dist}{Distance constant for the exponential defining the region for peak detection}

\item{frac.max}{Maximal fraction of gene end region for peal detection}

\item{frac.min}{Minimal fraction of gene end region for peal detection}
}
\value{
A bed6 file with TTS estimates incorporated, original TTSs are present if the TTS could not be estimated.
The output is a list including the bed frame along with several metrics (see example below and vignette).
}
\description{
Given regions within which to search for TTSs, we opperationally define the TTSs by binning the TTS search
regions, counting reads within the bins, fitting smooth spline curves to the bin counts, and
detecting points at which the curves peak and then decay towards zero. We applied the
constraint that there must be a specified number of bases in the TTS search region, otherwise the
TTS analysis is not applied. Similarly, if the number of knots identified is too low, then we set the
number of knots to a specified threshold. For this analysis, we set a peak search region at the beginning of
the TTS search region and identify the maximal peak from the spline fit. Then we identify the point
at which the spline fit decays to a threshold relative to the peak.
We reasoned that the peak search region should be largest, as a fraction of the TTS search region, for genes with
the greatest numbers of clipped bases. Such cases occur when the conventional
gene ends are proximal to identified TSSs, and we should include these entire regions for analysis of
the TTS. Similarly, we reasoned that for genes with substantially less clipped bases, and
correspondingly larger TTS search regions with greater potential for observing enhancers or divergent transcripts,
the peak search regions should be smaller fractions of the TTS search regions. We use
an exponential model to define the peak search regions.
The user should use the output of get.end.intervals() as an input to this function.
Note that gene ends will be set to zeros if there are no counts or if the interval is too small. 
The gene ends in tss will be the default if there are problems with estimating the TTS.
This function calls get.gene.end().
}
\examples{
# identify gene ends
add.to.end = max(bed.for.tts.eval$xy)
knot.div = 40
pk.thresh = 0.05
bp.bin = 50
knot.thresh = 5
cnt.thresh = 5
tau.dist = 10000
frac.max = 1
frac.min = 0.2
inferred.coords = get.TTS(bed=bed.for.tts.eval, tss=TSS.gene.filtered3,
                          bw.plus=bw.plus, bw.minus=bw.minus,
                          bp.bin=bp.bin, add.to.end=add.to.end,
                          knot.div=knot.div,
                          pk.thresh=pk.thresh, knot.thresh=knot.thresh,
                          cnt.thresh=cnt.thresh, tau.dist=tau.dist,
                          frac.max=frac.max, frac.min=frac.min)

}
