% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{TSS.count.dist}
\alias{TSS.count.dist}
\title{Get distances from identified transcription start sites (TSSs) to the nearest regions with peak reads}
\usage{
TSS.count.dist(bed = NULL, bw.plus = NULL, bw.minus = NULL,
  window = NULL, bp.bin = NULL)
}
\arguments{
\item{bed}{bed frame for TSS evaluation}

\item{bw.plus}{= plus strand bigWig data}

\item{bw.minus}{minus strand bigWig data}

\item{window}{region size, centered on the TSS for analysis}

\item{bp.bin}{the interval will be separated into adjacent bins of this size}
}
\value{
A list with three vector elements: raw, scaled, and dist.
All outputs are organized based on TSS position (left-upstream).
raw = binned raw counts. scaled = binned scaled counts (0,1).
dist = upper bound distances ( min(|dists|) = bp.bin ).
See examples and vignette for more details.
}
\description{
This function was designed to evaluate the results of out TSS identification analysis. We specify a region centered
on the TSS. We obtain read counts in bins that span this window. We sort the genes based on the bin with the
maximal reads and we scale the data to the interval (0,1) for visualization. We compute the distances between the TSS and
the bin with the max reads within the specified window.
}
\examples{
see eval.tss()
}
