% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{TTS.boundary.match}
\alias{TTS.boundary.match}
\title{Check for the fraction of identified TTSs that match the search boundry}
\usage{
TTS.boundary.match(coords = NULL, bed.for.tts.eval = NULL)
}
\arguments{
\item{coords}{bed6 frame with inferred coordinates}

\item{bed.for.tts.eval}{bed6 frame with TTS search regions (see get.end.intervals())}
}
\value{
A fraction [0,1].
}
\description{
This function considers information in both coords and bed.for.tts.eval (see vignette)
to identify the fraction of TTSs that match the boundary of the search region.
}
\examples{
frac.match = TTS.boundary.match(coords=coords, bed.for.tts.eval=bed.for.tts.eval)
}
