% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_ann.R
\name{tts.plot}
\alias{tts.plot}
\title{Plot inferred and largest interval coordinates}
\usage{
tts.plot(coords = NULL, gene.end = NULL, long.gene = NULL,
  bw.plus = NULL, bw.minus = NULL, bp.bin = 5, gene = NULL,
  xper = 0.2, yper = 1.3, add.to.end = NULL, tau.dist = NULL,
  frac.max = NULL, frac.min = NULL)
}
\arguments{
\item{coords}{A bed6 frame with inferred coordinates}

\item{gene.end}{A bed6 frame with peak search regions}

\item{long.gene}{A bed6 frame with largest interval coordinates}

\item{bw.plus}{Plus strand bigWig data}

\item{bw.minus}{Minus strand bigWig data}

\item{gene}{A gene for plotting}

\item{xper}{Fraction of the x-axis before the gene starts}

\item{yper}{Fraction of the y-axis with open space above the read data}

\item{add.to.end}{The maximal length of the search region (bp)}

\item{tau.dist}{Distance constant for the exponential defining the region for peak detection}

\item{frac.max}{Maximal fraction of gene end region for peak detection}

\item{frac.min}{Minimal fraction of gene end region for peak detection}
}
\value{
A plot with reads, the interred annotation (black), the largest interval annotation (gray),
the TTS search region (solid, vertical), and the TTS peak region (dashed, vertical).
}
\description{
This function plots the inferred and largest interval coordinates along with read data.
The plot annotates the peak search region and the total search region.
}
\examples{
# first run inferred.coords = get.TTS(...)
coords = inferred.coords$bed
gene.end = bed.for.tts.eval
long.gene = largest.interval.bed
tts.plot(coords=coords, gene.end=gene.end, long.gene=long.gene,
         gene = "Pparg", xper=0.2, yper=0.3,
         frac.min=frac.min, frac.max=frac.max,
         bw.plus=bw.plus, bw.minus=bw.minus, bp.bin=5,
         add.to.end=add.to.end, tau.dist=tau.dist)

}
